[KYUUBI #2471] add dynamic jar generator
