+++ kyuubi-common/src/test/scala/org/apache/kyuubi/operation/SparkQueryTests.scala
@@ -17,13 +17,10 @@
 
 package org.apache.kyuubi.operation
 
-import java.io.{File, FileOutputStream}
-import java.nio.file.Files
 import java.sql.{Date, SQLException, SQLTimeoutException, Timestamp}
-import java.util.jar.{JarEntry, JarOutputStream}
+import java.util.UUID
 
 import scala.collection.JavaConverters._
-import scala.tools.nsc._
 
 import org.apache.commons.io.FileUtils
 import org.apache.commons.lang3.StringUtils
@@ -628,17 +625,14 @@ trait SparkQueryTests extends HiveJDBCTestHelper {
           |
           |""".stripMargin
 
-      val jarFile = createJarFile(udfCode, "test-function.jar", jarDir.toString)
+      val jarFile = UserJarTestUtils.createJarFile(
+        udfCode,
+        "test",
+        s"test-function-${UUID.randomUUID}.jar",
+        jarDir.toString)
       val jarBytes = FileUtils.readFileToByteArray(jarFile)
       val jarStr = new String(java.util.Base64.getEncoder().encode(jarBytes))
-
-      // scalastyle:off
-      println(jarStr)
-      // magic string is base64 encoding of test-function.jar generated by new String(java.util.Base64.getEncoder().encode(jarBytes))
-      val jarStrMagic =
-        "UEsDBBQACAgIAPOpq1QAAAAAAAAAAAAAAAAVAAQAdGVzdC91dGlscy9NYXRoLmNsYXNz/soAAF2Qz27TQBDGv7Gzdp3m3wYCJBVFLRVqQKplqRzaCCEnpZFRQuuEIoHEYRMvqYtjq8ka0QNvxIlDDxx4AB4KsSni0pVGM9/sTzOa7/efn78A7KNFqCq5VG6u4mTpDoU6t0GE2oX4ItxEpDP3ZHIhp8qGSSiu0L3lVCSCsDm4KdyF/JRowB2v1DiepULlC9khsMmV5gnXFn07CKlrOkbbC+lI5+d+SMfEjDM/9Ff5o+NvvfXy8CkXokvW59Ggnz60DM62fPfDYW4TbYseVcwG/8e94dwvOOajIObELOMxa3vZe2/ee8I9kxm2eTjYfdbaN8qXr2d7G52XbRZy4XsyGDUp6lc5exHUu1ThrBzIyiazDspUpj4Zl9ttEZEgjzm8RjBFFOkzdoOgHRDoq44rbc0tv3ZslAn28OTobPBqh1Af3AY6JVRRc1ABL8FBsairO4RCL4uknjeU6jyLTsVCzKWSC+1YcZzli6k8jhP93xzlqYrn8l28jCeJ9NM0U0LFWarBtf+mw4OBAlaP9Aq9ROuGVg2YugKcH6i3Nq5x9/sNdG/VMrCOGko6TNzXDQMMD3Re1zMsHfYSa2jqhsmAv1BLBwj7YSlmvwEAADMCAABQSwMEFAAICAgA86mrVAAAAAAAAAAAAAAAABYAAAB0ZXN0L3V0aWxzL01hdGgkLmNsYXNzXVHLTsJAFD23LZQWFATfj8SFC2EhMTFxIXHhK2kCauJj7QCjlJQ2KYPRf3JDXJBo4gf4UcY7A3Fhm5lzH2fuOZ1+/3x8ATjAFqGk5FDVRyqMhvWWUL0dF8TVvngW9UjET/Wrdl92lAub4Gvu3rAjIkFwW1dnd83zHUK5+X/GESHX6ERhHKpjgr1bvSdkGyYtwIfnw0KeG6LbJWR2g6AaEOiF1yvBUb1wyGHgw0G+gCxcj/mL3DlNupJQbIaxvBwN2jK9Fe1IagcJm7oXaajzWbHUkqqXdK9FKgZSyZRn+jfJKO3Ii1D3izf6Q4KYbcogfkzYialgn9Ucvh+Cp8U5sjnK8SIsc1Zh1I/7CWuCwtgkhDnM80FNOGG0GZ2NzYc3RgsrvGfNsQxWed+e9rFgxuuoiLLpOyhxZGGNY8/icsm8lr62mf4ho6Z6tQkqtXcsjf809MWyMTbrGJ3ClDjTsbFumBlsMM7zfzZjjXXCpjbyC1BLBwid5JsCcgEAABsCAABQSwECFAAUAAgICADzqatU+2EpZr8BAAAzAgAAFQAEAAAAAAAAAAAAAAAAAAAAdGVzdC91dGlscy9NYXRoLmNsYXNz/soAAFBLAQIUABQACAgIAPOpq1Sd5JsCcgEAABsCAAAWAAAAAAAAAAAAAAAAAAYCAAB0ZXN0L3V0aWxzL01hdGgkLmNsYXNzUEsFBgAAAAACAAIAiwAAALwDAAAAAA=="
-      // scalastyle:on
-      val currentTime = System.nanoTime()
+      val jarName = s"test-function-${UUID.randomUUID}.jar"
 
       val code0 = """spark.sql("SET kyuubi.operation.language").show(false)"""
       statement.execute(code0)
@@ -648,9 +642,9 @@ trait SparkQueryTests extends HiveJDBCTestHelper {
         s"""
            |import java.io.{BufferedOutputStream, File, FileOutputStream}
            |val dir = spark.sparkContext.getConf.get("spark.repl.class.outputDir")
-           |val jarFile = new File(dir, "test-function-$currentTime.jar")
+           |val jarFile = new File(dir, "$jarName")
            |val bos = new BufferedOutputStream(new FileOutputStream(jarFile))
-           |val path = "$jarStrMagic"
+           |val path = "$jarStr"
            |bos.write(java.util.Base64.getDecoder.decode(path))
            |bos.close()
            |val jarPath = jarFile.getAbsolutePath
@@ -670,42 +664,6 @@ trait SparkQueryTests extends HiveJDBCTestHelper {
       val rs = statement.executeQuery(code2)
       rs.next()
       assert(rs.getString(1) == "x: Int = 3")
-    }
-  }
-
-  def createJarFile(codeText: String, jarName: String, outputDir: String): File = {
-    val codeFile = new File(outputDir, "test.scala")
-    FileUtils.writeStringToFile(codeFile, codeText, "UTF-8")
-
-    val settings = new Settings
-    settings.outputDirs.setSingleOutput(outputDir)
-    settings.usejavacp.value = true
-    val global = new Global(settings)
-    val runner = new global.Run
-    runner.compile(List(codeFile.getAbsolutePath))
-    val jarFile = new File(outputDir, jarName)
-    val targetJar = new JarOutputStream(new FileOutputStream(jarFile))
-    add(outputDir + "/test", targetJar, outputDir + "/")
-    targetJar.close
-    jarFile.createNewFile()
-    jarFile
-  }
-
-  private def add(folder: String, target: JarOutputStream, replacement: String): Unit = {
-    val source = new File(folder)
-    if (source.isDirectory) {
-      for (nestedFile <- source.listFiles) {
-        add(nestedFile.getAbsolutePath, target, replacement)
-      }
-    } else {
-      val entry = new JarEntry(source.getPath
-        .replace("\\", "/")
-        .replace(replacement, ""))
-      entry.setTime(source.lastModified)
-      target.putNextEntry(entry)
-      val byteArray = Files.readAllBytes(source.toPath)
-      target.write(byteArray)
-      target.closeEntry()
     }
   }
